:sparkles: Make field configuration compatible with update forms — #349
