+++ src/sdg/producten/models/mixins.py
@@ -4,25 +4,17 @@ from django.db import models
 from django.utils.translation import ugettext_lazy as _
 
 from sdg.core.constants import TaalChoices
-from sdg.core.models import ProductFieldConfiguration
+from sdg.core.models.mixins import FieldConfigurationMixin
 from sdg.producten.types import ProductFieldInfo
 
 
-class ProductFieldMixin:
-    _configuration: ProductFieldConfiguration = None
-
-    @property
-    def configuration(self):
-        if self._configuration is None:
-            self._configuration = ProductFieldConfiguration.get_solo()
-        return self._configuration
-
+class ProductFieldMixin(FieldConfigurationMixin):
     def _get_field(self, field) -> ProductFieldInfo:
         """Gets field specific information for products."""
         if isinstance(field, str):
             field = self._meta.get_field(field)
 
-        value, is_reference = self._get_field_value(field)
+        value, is_reference = self.get_field_value(field)
 
         return ProductFieldInfo(
             name=field.name,
@@ -31,7 +23,7 @@ class ProductFieldMixin:
             help_text=field.help_text,
             is_reference=is_reference,
             type=field.get_internal_type(),
-            configuration=self.configuration.get_field(
+            configuration=self.configuration.for_field(
                 prefix=self._meta.model_name,
                 name=field.name,
             ),
